# apps/server/generated/@typespec/openapi3/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expense service",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/v1/transactions": {
      "get": {
        "operationId": "Transactions_list",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "explode": false
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "explode": false
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "Transactions_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/summary": {
      "get": {
        "operationId": "Transactions_summary",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "explode": false
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSummary"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/v1/transactions/{transactionId}": {
      "get": {
        "operationId": "Transactions_get",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "Transactions_update",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "Transactions_delete",
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "InternalServerError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INTERNAL_SERVER_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number",
            "enum": [
              500
            ]
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "NOT_FOUND"
            ]
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number",
            "enum": [
              404
            ]
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
...
      "Transaction": {
        "type": "object",
        "required": [
          "id",
          "accountId",
          "amount",
          "categoryId",
          "userId",
          "type",
          "date"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "accountId": {
            "type": "integer",
            "minimum": 1
          },
          "amount": {
            "type": "number",
            "format": "double",
            "minimum": 0.01
          },
          "categoryId": {
            "type": "integer",
            "minimum": 1
          },
          "userId": {
            "type": "integer",
            "minimum": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
      "TransactionCreate": {
        "type": "object",
        "required": [
          "amount",
          "categoryId",
          "type"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "minimum": 0.01
          },
          "categoryId": {
            "type": "integer",
            "minimum": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
      "TransactionSummary": {
        "type": "object",
        "required": [
          "incomeTotal",
          "expenseTotal",
          "periodBalance",
          "totalBalance"
        ],
        "properties": {
          "incomeTotal": {
            "type": "number",
            "format": "double"
          },
          "expenseTotal": {
            "type": "number",
            "format": "double"
          },
          "periodBalance": {
            "type": "number",
            "format": "double"
          },
          "totalBalance": {
            "type": "number",
            "format": "double"
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
      "TransactionUpdate": {
        "type": "object",
        "required": [
          "amount",
          "categoryId",
          "type",
          "date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "minimum": 0.01
          },
          "categoryId": {
            "type": "integer",
            "minimum": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
...
      "ValidationError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number",
            "enum": [
              400
            ]
          }
        },
        "unevaluatedProperties": {
          "not": {}
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Single server endpoint",
      "variables": {}
    }
  ]
}